# Reduce Swift compile time.